Protoss bot
added a basic code that runs a protoss ai- protoss are NOT zerg, and you
can't just go for a straight rush without building up your base a
little, so this bot will always lose. But, it runs. Also builds pylons
when supply is needed, and builds a gateway next to the pylon. long ways
to go.